requirements.txt
# AI Agent Dependencies - Optimized for MacBook Air M4
# Python 3.11+ required

# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4

# Local LLM - Metal-optimized for Apple Silicon
llama-cpp-python==0.3.4

# Vector Memory Store - Lightweight alternative to FAISS
chromadb==0.5.23

# Code Analysis
ast-grep-py==0.31.1

# HTTP Client for web verification
httpx==0.28.1

# Utilities
python-multipart==0.0.18
orjson==3.10.12

# Security scanning (optional but recommended)
bandit==1.8.0
